+++ src/pages/diagnosis/answerButtons/defButton/index.style.tsx
