+++ src/pages/diagnosis/answerButtons/etcButton/index.style.tsx
