+++ src/pages/diagnosis/answerButtons/nextButton/index.style.tsx
