+++ src/pages/diagnosis/answerButtons/nextButton/index.tsx
