+++ src/pages/diagnosis/answerButtons/numberButtons/alcoholButton/index.style.tsx
