+++ src/pages/diagnosis/answerButtons/numberButtons/alcoholButton/index.tsx
