+++ src/pages/diagnosis/answerButtons/numberButtons/countButton/index.style.tsx
