+++ src/pages/diagnosis/answerButtons/numberButtons/countButton/index.tsx
