+++ src/pages/diagnosis/answerButtons/numberButtons/durationButton/index.style.tsx
