+++ src/pages/diagnosis/answerButtons/numberButtons/index.tsx
