+++ src/pages/diagnosis/answerButtons/numberButtons/previousTimeButton/index.style.tsx
