+++ src/pages/diagnosis/answerButtons/numberButtons/smockingButton/index.style.tsx
