+++ src/pages/diagnosis/answerButtons/stringButton/index.style.tsx
